feat: add code for typescript 2021 day 16
